requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.31.0
black==24.8.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
et_xmlfile==2.0.0
fastapi==0.126.0
fastapi-cprofile==0.0.2
fastjsonschema==2.16.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
jwt==1.3.1
MouseInfo==0.1.3
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
pytest==8.3.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.6
pytweening==1.2.0
pytz==2025.2
requests==2.31.0
rsa==4.9.1
rubicon-objc==0.4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.2.2
uvicorn==0.40.0
alembic==1.11.1
passlib==1.7.4
bcrypt-5.0.0
